requirements.txt
discord.py
yt-dlp
ffmpeg
python-dotenv
